Added display to owner label
Also removed drawKeyPoints from map2d

+++ spring2016/Navi/navi_app/app/src/main/java/edu/stanford/navi/OwnerMapActivity.java
@@ -299,7 +299,7 @@ public class OwnerMapActivity extends BaseActivity implements View.OnClickListen
     public void setUpMap() {
         android.graphics.Point screenSize = new android.graphics.Point();
         getWindowManager().getDefaultDisplay().getSize(screenSize);
-        map = new Map2D(this, (int) screenSize.x, (int)screenSize.y);
+        map = new Map2D(this, screenSize.x, screenSize.y);
         mapBitmap = map.imgBmp.copy(Bitmap.Config.ARGB_8888, true);
         imageView = (ImageView) findViewById(R.id.ownerMap);
         imageView.setImageBitmap(mapBitmap);
